# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Pharmacy Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SupplierWithBalance" }
    post:
      operationId: createSupplier
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Supplier" }
  /suppliers/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getSupplier
      tags: [suppliers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SupplierDetail" }
  /purchase-orders:
    post:
      operationId: createPurchaseOrder
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PurchaseOrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrder" }
  /purchase-orders/{id}/payments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    post:
      operationId: payPurchaseOrder
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SupplierPaymentInput" }
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrder" }

  # ─── Customers + Udhari ────────────────────────────────────
...
  /sales:
    get:
      operationId: listSales
      tags: [sales]
      parameters:
        - { name: from, in: query, schema: { type: string, format: date } }
        - { name: to, in: query, schema: { type: string, format: date } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SaleSummary" }
    post:
      operationId: createSale
      tags: [sales]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SaleInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }
  /sales/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      operationId: getSale
      tags: [sales]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Sale" }

  # ─── Analytics ─────────────────────────────────────────────
...
components:
  schemas:
...
    Supplier:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        name: { type: string }
        contactPhone: { type: ["string", "null"] }
        gstin: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        creditDays: { type: integer }
    SupplierWithBalance:
      allOf:
        - $ref: "#/components/schemas/Supplier"
        - type: object
          required: [outstandingMinor, totalPurchasedMinor, invoiceCount]
          properties:
            outstandingMinor: { type: integer }
            totalPurchasedMinor: { type: integer }
            invoiceCount: { type: integer }
            nextDueDate: { type: ["string", "null"], format: date }
    SupplierInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        contactPhone: { type: ["string", "null"] }
        gstin: { type: ["string", "null"] }
        address: { type: ["string", "null"] }
        creditDays: { type: integer, default: 30 }
    SupplierDetail:
      allOf:
        - $ref: "#/components/schemas/SupplierWithBalance"
        - type: object
          properties:
            invoices:
              type: array
              items: { $ref: "#/components/schemas/PurchaseOrder" }
    PurchaseOrder:
      type: object
      required:
        - id
        - supplierId
        - supplierName
        - invoiceNumber
        - invoiceDate
        - totalMinor
        - paidMinor
        - status
      properties:
        id: { type: string }
        supplierId: { type: string }
        supplierName: { type: string }
        invoiceNumber: { type: string }
        invoiceDate: { type: string, format: date }
        dueDate: { type: ["string", "null"], format: date }
        totalMinor: { type: integer }
        paidMinor: { type: integer }
        status: { type: string, enum: [pending, partial, paid] }
        notes: { type: ["string", "null"] }
    PurchaseOrderLineInput:
      type: object
      required: [medicineName, batchNumber, expiryDate, qty, buyPriceMinor, sellPriceMinor]
      properties:
        medicineId: { type: ["string", "null"] }
        medicineName: { type: string }
        batchNumber: { type: string }
        expiryDate: { type: string, format: date }
        qty: { type: integer }
        buyPriceMinor: { type: integer }
        sellPriceMinor: { type: integer }
    PurchaseOrderInput:
      type: object
      required: [supplierId, invoiceNumber, invoiceDate, lines]
      properties:
        supplierId: { type: string }
        invoiceNumber: { type: string }
        invoiceDate: { type: string, format: date }
        dueDate: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
        lines:
          type: array
          items: { $ref: "#/components/schemas/PurchaseOrderLineInput" }
    SupplierPaymentInput:
      type: object
      required: [amountMinor, method]
      properties:
        amountMinor: { type: integer }
        method: { type: string, enum: [cash, upi, bank, cheque] }
        reference: { type: ["string", "null"] }

    # ─── Customers / Udhari ─────
...
    SaleLineInput:
      type: object
      required: [batchId, qty, sellPriceMinor]
      properties:
        batchId: { type: string }
        qty: { type: integer }
        sellPriceMinor: { type: integer }
        discountMinor: { type: integer, default: 0 }
    SaleLine:
      type: object
      required: [id, batchId, medicineName, qty, sellPriceMinor, lineTotalMinor]
      properties:
        id: { type: string }
        batchId: { type: string }
        medicineName: { type: string }
        batchNumber: { type: string }
        qty: { type: integer }
        sellPriceMinor: { type: integer }
        buyPriceMinor: { type: integer }
        discountMinor: { type: integer }
        lineTotalMinor: { type: integer }
    SaleInput:
      type: object
      required: [lines, paymentMethod]
      properties:
        customerId: { type: ["string", "null"] }
        doctorId: { type: ["string", "null"] }
        prescriptionNo: { type: ["string", "null"] }
        discountMinor: { type: integer, default: 0 }
        paymentMethod: { type: string, enum: [cash, upi, card, udhari] }
        paidMinor: { type: integer }
        lines:
          type: array
          items: { $ref: "#/components/schemas/SaleLineInput" }
    SaleSummary:
      type: object
      required: [id, invoiceNo, totalMinor, profitMinor, paymentMethod, soldAt]
      properties:
        id: { type: string }
        invoiceNo: { type: string }
        customerName: { type: ["string", "null"] }
        doctorName: { type: ["string", "null"] }
        totalMinor: { type: integer }
        profitMinor: { type: integer }
        paymentMethod: { type: string }
        soldAt: { type: string, format: date-time }
    Sale:
      allOf:
        - $ref: "#/components/schemas/SaleSummary"
        - type: object
          required: [lines, subtotalMinor, taxMinor, discountMinor, paidMinor]
          properties:
            customerId: { type: ["string", "null"] }
            doctorId: { type: ["string", "null"] }
            prescriptionNo: { type: ["string", "null"] }
            subtotalMinor: { type: integer }
            taxMinor: { type: integer }
            discountMinor: { type: integer }
            paidMinor: { type: integer }
            costMinor: { type: integer }
            lines:
              type: array
              items: { $ref: "#/components/schemas/SaleLine" }

    # ─── Analytics ─────
...