# lib/api-zod/src/generated/types/purchaseOrderLineInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface PurchaseOrderLineInput {
  /** @nullable */
  medicineId?: string | null;
  medicineName: string;
  batchNumber: string;
  expiryDate: Date;
  qty: number;
  buyPriceMinor: number;
  sellPriceMinor: number;
}