# lib/api-zod/src/generated/types/saleLineInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface SaleLineInput {
  batchId: string;
  qty: number;
  sellPriceMinor: number;
  discountMinor?: number;
}