# lib/api-zod/src/generated/types/supplierPaymentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Pharmacy Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { SupplierPaymentInputMethod } from "./supplierPaymentInputMethod";

export interface SupplierPaymentInput {
  amountMinor: number;
  method: SupplierPaymentInputMethod;
  /** @nullable */
  reference?: string | null;
}